backend/requirements.txt
# ================================
# Core Framework & Server
# ================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiofiles==24.1.0

# ================================
# Pydantic & Validation
# ================================
pydantic==2.10.3
pydantic[email]==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1

# ================================
# RAG System & Embeddings
# ================================
sentence-transformers>=2.3.0
huggingface_hub>=0.25.0
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
torch>=2.0.0
transformers>=4.30.0

# ================================
# Google Gemini AI
# ================================
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-auth==2.36.0

# ================================
# Database & Supabase
# ================================
supabase>=2.8.0
postgrest>=0.16.0
gotrue>=2.8.0
realtime>=2.0.0
storage3>=0.7.0
httpx>=0.26,<0.28

# ================================
# Authentication & Security
# ================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyJWT==2.10.1
cryptography==43.0.3

# ================================
# OAuth & Social Auth
# ================================
authlib==1.3.2
requests==2.32.3
oauthlib==3.2.2

# ================================
# Document Processing
# ================================
pdfplumber==0.11.4
pytesseract==0.3.13
pillow==11.0.0
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
tabulate==0.9.0

# ================================
# OCR & Image Processing
# ================================
opencv-python==4.10.0.84
pdf2image==1.17.0

# ================================
# Environment & Configuration
# ================================
python-dotenv==1.0.1
pyyaml==6.0.2

# ================================
# Data Processing & Utilities
# ================================
python-dateutil==2.9.0.post0
typing-extensions==4.12.2
tqdm==4.67.1
regex==2024.11.6

# ================================
# CORS & HTTP
# ================================
httpcore==1.0.7
h11==0.14.0
anyio==4.7.0

# ================================
# Logging & Monitoring
# ================================
loguru==0.7.2
rich==13.9.4

# ================================
# Testing
# ================================
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-httpx==0.30.0 

# ================================
# Development Tools
# ================================
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# ================================
# Additional ML/AI Tools
# ================================
langchain==0.3.10
langchain-community==0.3.10
chromadb==0.5.23
tiktoken==0.8.0

# ================================
# Job Queue & Background Tasks
# ================================
celery==5.4.0
redis==5.2.1

# ================================
# WebSockets & Real-time
# ================================
websockets==14.1
python-socketio==5.12.0